pyproject.toml
[project]
name = "geo-content-platform"
version = "3.2.0"
description = "Multi-Agent GEO Content Optimization Platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Tocanan.ai" }
]
keywords = ["geo", "seo", "content-optimization", "multi-agent", "llm", "openai", "anthropic"]

dependencies = [
    # OpenAI Agent SDK with tracing
    "openai-agents>=0.1.0",
    "openai>=1.50.0",

    # Anthropic SDK for Claude
    "anthropic>=0.39.0",

    # Pathway for web harvesting
    "pathway>=0.14.0",

    # Language detection
    "langdetect>=1.0.9",
    "arabic-reshaper>=3.0.0",
    "python-bidi>=0.4.2",

    # Web scraping
    "beautifulsoup4>=4.12.0",
    "httpx>=0.27.0",
    "playwright>=1.40.0",

    # Document processing
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",

    # Data handling
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",

    # API
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "tenacity>=8.2.0",

    # AWS
    "boto3>=1.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/geo_content"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
